+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * BREAKING: Added properties `onTapMarker`, `onTapInfoWindow`, `infoWindowTitle` and `isVisible` to `MarkerIconInfo`, and removed corresponding params for source, destination, driver from `GoogleMapsWidget`
 * Changed internal implementation of the widget
 * Added `layoutDirection` property
+* Added `onPolylineUpdate` callback
 * Exposed state class to allow updating source/destination lat lng, or interacting with google maps con directly
 * Updated a dependency to the latest release
 * Updated example app